Restructure scvi docker image
