+++ docker/scvi/scripts/main/annotate_cells.py
