+++ docker/scvi/scripts/main/clustering_umap.py
