+++ docker/scvi/scripts/main/process.py
@@ -1,7 +1,7 @@
 import argparse
 import pandas as pd
 import scanpy
-from util.helpers import score_cell_cycle
+from utility.helpers import score_cell_cycle
 
 
 parser = argparse.ArgumentParser(
